# new-iVan-master/docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swipped API
  version: 1.0.0
  description: |
    ## Introduction
    Unified API documentation for Swipped.

    This document groups endpoints by business domain so API consumers can navigate
    authentication, public discovery, and protected/admin operations more easily.

    ## Base URL
    - Live: `https://dev.swippped.com`
    - Local: `http://localhost:3001`
servers:
  - url: https://dev.swippped.com
    description: Live domain
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/customer/jobs/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Customer]
      summary: "Customer - Jobs: Get job details"
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses: *stdResponses
  /api/customer/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Customer]
      summary: "Customer - Orders: Get order details"
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses: *stdResponses
  /api/customer/orders/{id}/tip:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Customer]
      summary: "Customer - Orders: Add tip"
      requestBody: *defaultBody
      responses: *stdResponses
...
  /api/feedbacks/pending:
    get: *defaultGet
  /api/feedbacks/submit:
    post: *customerPost
...
components:
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    SearchParam:
      name: search
      in: query
      required: false
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1